+++ src/main/java/step2/domain/BridgeBuildStrategy.java
@@ -1,5 +1,6 @@
 package step2.domain;
 
+@FunctionalInterface
 public interface BridgeBuildStrategy {
     boolean canBuild();
 }
