+++ src/main/java/step2/domain/Ladder.java
@@ -3,10 +3,10 @@ package step2.domain;
 import java.util.ArrayList;
 
 public class Ladder {
-    private int height;
-    private int width;
+    private final int height;
+    private final int width;
 
-    private ArrayList<Line> lines = new ArrayList<>();
+    private final ArrayList<Line> lines = new ArrayList<>();
 
     public Ladder(int height, int width) {
         this.height = height;
@@ -28,8 +28,23 @@ public class Ladder {
         return lines;
     }
 
-    public Ladder apply(ArrayList<Line> lines) {
-        this.lines = lines;
+    public Ladder buildBridges(BridgeBuildStrategy bridgeBuildStrategy) {
+        for (int lineNumber = 0; lineNumber < width - 1; lineNumber++) {
+            buildBridge(lineNumber, height, bridgeBuildStrategy);
+        }
         return this;
+    }
+
+    private ArrayList<Line> buildBridge(int lineNumber, int height, BridgeBuildStrategy bridgeBuildStrategy) {
+        for (int position = 0; position < height; position++) {
+            if (areLinesOverlapping(lines, lineNumber, position) || !bridgeBuildStrategy.canBuild()) continue;
+            lines.get(lineNumber).addBridge(position);
+        }
+        return lines;
+    }
+
+    private boolean areLinesOverlapping(ArrayList<Line> lines, int lineNumber, int position) {
+        if (lineNumber - 1 < 0) return false;
+        return lines.get(lineNumber - 1).isPresent(position);
     }
 }
