--- src/main/java/step2/domain/RandomBridgeBuildStrategy.java
@@ -1,9 +1,0 @@
-package step2.domain;
-
-public class RandomBridgeBuildStrategy implements  BridgeBuildStrategy {
-
-    @Override
-    public boolean canBuild() {
-        return Math.random() < 0.5;
-    }
-}
